+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to
 
 ### Added
 
-- ✨(backend) add a is_first_connection flag to the User model#1938
+- ✨(backend) add a is_first_connection flag to the User model #1938
+- ✨(backend) add resource server api #1923
 
 ## [v4.7.0] - 2026-03-09